+++ CHANGELOG.rst
@@ -9,6 +9,8 @@ This document describes changes between each past release.
 **Bug fixes**
 
 - Relax content-type validation when no body is posted (fixes #507)
+- Fix creation events not sent for implicit creation of objects in the ``default``
+  bucket (fixes #529)
 
 **New features**